bug 6381 fixed: Refresh the statistical information section automatically when run a catalog or schema analysis
git-svn-id: http://talendforge.org/svn/top/trunk@21996 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.dataprofiler.core/src/org/talend/dataprofiler/core/ui/editor/preview/HideSeriesPanel.java
@@ -172,7 +172,7 @@ public class HideSeriesPanel extends JPanel implements ActionListener {
             XYPlot plot = chart.getXYPlot();
             xyRenderer = plot.getRenderer();
             NumberAxis rangeAxis = (NumberAxis) plot.getRangeAxis();
-            int minYValue = (int) getMinYValue(plot, true);
+            double minYValue = getMinYValue(plot, true);
             rangeAxis.setAutoRange(false);
             rangeAxis.setRange(0, minYValue * 100);
             rangeAxis.setTickUnit(new NumberTickUnit(minYValue * 5));
